add IT to check the DB lock mechanism
This test is ignore since it's supposed to work only locally
